Added cygwin support
git-svn-id: svn://svn.code.sf.net/p/aceos/code@4 78f6c469-4e3a-0410-bca2-dfd793a85426

+++ img/copy_file_to_image.sh
@@ -1,14 +1,27 @@
 #! /bin/sh
+USAGE="Usage: `basename $0` <image_file> <local_file> <image_path>"
+if [ $# -lt 3 ] ; then
+	echo "$USAGE"
+	exit 1
+fi
 
-free_loop=`losetup -f`
-losetup $free_loop $1
+if [ $OSTYPE = cygwin ] ; then
+	WINDOWS_PATH=`cygpath -a -w $1`
+	filedisk /mount 0 $WINDOWS_PATH m:
+	cp $2 m:/$3
+	filedisk /umount m:
+else
+	free_loop=`losetup -f`
+	losetup $free_loop $1
 
-mkdir $ACE_ROOT/temp
-mount $free_loop $ACE_ROOT/temp
+	mkdir $ACE_ROOT/temp
+	mount $free_loop $ACE_ROOT/temp
 
-cp $2 $ACE_ROOT/temp/$3
+	cp $2 $ACE_ROOT/temp/$3
 
-umount $ACE_ROOT/temp
-rm -rf $ACE_ROOT/temp
+	umount $ACE_ROOT/temp
+	rm -rf $ACE_ROOT/temp
 
-losetup -d $free_loop
+	losetup -d $free_loop
+fi
+
